+++ lib/punchblock/translator/asterisk.rb
@@ -70,9 +70,16 @@ module Punchblock
           register_call call
         end
 
-        if event['Channel'] && call = call_for_channel(event['Channel'])
-          pb_logger.trace "Found call by channel matching this event. Sending to call #{call.id}"
-          call.process_ami_event! event
+        if (event['Channel'] && call_for_channel(event['Channel'])) ||
+            (event['Channel1'] && call_for_channel(event['Channel1'])) ||
+            (event['Channel2'] && call_for_channel(event['Channel2']))
+          [event['Channel'], event['Channel1'], event['Channel2']].compact.each do |channel|
+            call = call_for_channel channel
+            if call
+              pb_logger.trace "Found call by channel matching this event. Sending to call #{call.id}"
+              call.process_ami_event! event
+            end
+          end
         elsif event.name.downcase == "asyncagi" && event['SubEvent'] == "Start"
           handle_async_agi_start_event event
         end