+++ spec/punchblock/translator/asterisk_spec.rb
@@ -458,8 +458,34 @@ module Punchblock
             call.expects(:process_ami_event!).once.with ami_event
             subject.handle_ami_event ami_event
           end
+
+          context 'with a Channel1 and Channel2 specified on the event' do
+            let :ami_event do
+              RubyAMI::Event.new('BridgeAction').tap do |e|
+                e['Privilege'] = "call,all"
+                e['Response'] = "Success"
+                e['Channel1']  = "SIP/1234-00000000"
+                e['Channel2']  = "SIP/5678-00000000"
+              end
+            end
+
+            context 'with calls for those channels' do
+              let(:call2) do
+                Asterisk::Call.new "SIP/5678-00000000", subject, "agi_request%3A%20async%0Aagi_channel%3A%20SIP%2F1234-00000000%0Aagi_language%3A%20en%0Aagi_type%3A%20SIP%0Aagi_uniqueid%3A%201320835995.0%0Aagi_version%3A%201.8.4.1%0Aagi_callerid%3A%205678%0Aagi_calleridname%3A%20Jane%20Smith%0Aagi_callingpres%3A%200%0Aagi_callingani2%3A%200%0Aagi_callington%3A%200%0Aagi_callingtns%3A%200%0Aagi_dnid%3A%201000%0Aagi_rdnis%3A%20unknown%0Aagi_context%3A%20default%0Aagi_extension%3A%201000%0Aagi_priority%3A%201%0Aagi_enhanced%3A%200.0%0Aagi_accountcode%3A%20%0Aagi_threadid%3A%204366221312%0A%0A"
+              end
+
+              before { subject.register_call call2 }
+
+              it 'should send the event to both calls and to the connection once as a PB event' do
+                subject.wrapped_object.expects(:handle_pb_event).once
+                call.expects(:process_ami_event!).once.with ami_event
+                call2.expects(:process_ami_event!).once.with ami_event
+                subject.handle_ami_event ami_event
+              end
+            end
+          end
         end
-      end
+      end#handle_ami_event
 
       describe '#send_ami_action' do
         it 'should send the action to the AMI client' do